+++ tomcat/src/main/java/org/apache/catalina/controller/Controller.java
@@ -1,0 +1,8 @@
+package org.apache.catalina.controller;
+
+import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.HttpResponse;
+
+public interface Controller {
+    void service(HttpRequest request, HttpResponse response) throws Exception;
+}
