+++ tomcat/src/main/java/org/apache/coyote/http11/HttpRequest.java
@@ -1,10 +1,8 @@
-package org.apache.catalina.http;
+package org.apache.coyote.http11;
 
 import org.apache.catalina.storage.Session;
 import org.apache.catalina.storage.SessionManager;
 import org.apache.catalina.storage.Cookies;
-import org.apache.catalina.http.vo.FileExtension;
-import org.apache.catalina.http.vo.Mime;
 
 import java.util.Collections;
 import java.util.Map;
