+++ tomcat/src/main/java/org/apache/coyote/http11/HttpStatus.java
@@ -1,4 +1,4 @@
-package org.apache.catalina.http.vo;
+package org.apache.coyote.http11;
 
 import java.util.HashMap;
 import java.util.Map;
