+++ tomcat/src/main/java/org/apache/coyote/http11/Path.java
@@ -1,6 +1,5 @@
-package org.apache.catalina.http;
+package org.apache.coyote.http11;
 
-import org.apache.catalina.http.vo.FileExtension;
 import org.apache.coyote.http11.util.HttpRequestParser;
 
 import java.util.Map;
